Added global conftest to download files
