+++ tests/test_wpa.py
@@ -1,9 +1,5 @@
 """Tests WPA module."""
-import json
-import os
-
 import pytest
-import requests
 
 from awpy.analytics.wpa import round_win_probability, state_win_probability
 from awpy.parser import DemoParser
@@ -18,11 +14,6 @@ class TestStates:
 
     def setup_class(self):
         """Setup class by instantiating parser."""
-        with open("tests/test_data.json", encoding="utf-8") as f:
-            self.demo_data = json.load(f)
-        self._get_demofile(
-            demo_link=self.demo_data["default"]["url"], demo_name="default"
-        )
         self.parser = DemoParser(demofile="default.dem", log=True, parse_rate=256)
         self.data = self.parser.parse()
 
@@ -30,24 +21,6 @@ class TestStates:
         """Set parser to none."""
         self.parser = None
         self.data = None
-        files_in_directory = os.listdir()
-        if filtered_files := [
-            file for file in files_in_directory if file.endswith((".dem", ".json"))
-        ]:
-            for f in filtered_files:
-                os.remove(f)
-
-    @staticmethod
-    def _get_demofile(demo_link: str, demo_name: str) -> None:
-        print(f"Requesting {demo_link}")
-        r = requests.get(demo_link, timeout=100)
-        with open(f"{demo_name}.dem", "wb") as demo_file:
-            demo_file.write(r.content)
-
-    @staticmethod
-    def _delete_demofile(demo_name: str) -> None:
-        print(f"Removing {demo_name}")
-        os.remove(f"{demo_name}.dem")
 
     def test_state_win_probability(self):
         """Tests state_win_probability."""
